+++ src/doomstat.c
@@ -189,6 +189,7 @@ int event_timers[NUMTIMERS];
 int extra_gibbing;
 int bloodier_gibbing;
 int zdoom_item_drops;
+int blood_splats; // [Cherry]
 
 // CFG only
 int chasecam_crosshair;